+++ internal/domain/game.go
@@ -1,0 +1,14 @@
+package domain
+
+import (
+	"time"
+)
+
+type Game struct {
+	ID          int       `json:"id"`
+	Title       string    `json:"title"`
+	Description string    `json:"description"`
+	Genre       string    `json:"genre"`
+	ReleaseDate time.Time `json:"releaseDate"`
+	Developer   string    `json:"developer"`
+}
