+++ internal/domain/user.go
@@ -1,0 +1,8 @@
+package domain
+
+type User struct {
+	ID           int    `json:"id"`
+	Username     string `json:"username"`
+	Email        string `json:"email"`
+	PasswordHash string `json:"passwordHash"`
+}
